+++ src/sybil_extras/parsers/abstract/grouped_source.py
@@ -314,7 +314,7 @@ class AbstractGroupedSourceParser:
         evaluator: Evaluator,
         directive: str,
         pad_groups: bool,
-        no_pad_separator_lines: int = 1,
+        no_pad_separator_lines: int = 2,
     ) -> None:
         """
         Args: