+++ src/sybil_extras/parsers/docutils_rst/grouped_source.py
@@ -51,5 +51,4 @@ class GroupedSourceParser(AbstractGroupedSourceParser):
             evaluator=evaluator,
             directive=directive,
             pad_groups=pad_groups,
-            no_pad_separator_lines=2,
         )